Extract BlogTimeline and Timeline to separate file
