+++ src/priviblur_extractor/models/__init__.py
@@ -1,3 +1,3 @@
-from . import base, timeline, blog, misc
+from . import base, timeline, timelines, misc
 
 from .base import VERSION
